completed the HTML and CSS file for this project
After creating the initial body layout for this webpage. I have gone ahead to fill out the entire page with information about the cycling websites.
I have also included cards which distinct the different sections that customers may pick from depending on the service they are looking for. At the bottom of the page are the testimonies for this company.